Feat: ViewMovie 부분 Component로 이동
